Added Glass bottom Boats
I had to move the texture over by 1 pixel as compared to the original due to the Vanilla game's UV mapping being 1 pixel off the colours of the sides of the boat in the texture file itself. Looks better in game than the original Glass Bottom Boats texture pack but if youre pedantic you'll still see the offset (but the offset is also in Vanilla so :shrug:).