Updated menu backgrounds for 1.21.8
This has been broken since 1.20.5, because the old cobble background file is still there (removed in pack format 28).

1. That cobble sprite had poorly managed colours so I've under- rather than overexposed it
2. Made a half-arsed attempt at making it look nice with black borders and stuff so list backgrounds (the darker parts) don't look HORRID. The list background is actually the default texture, but it's included since Mojang added a different list background texture to Programmer Art that includes Dirt and I want a transparent dimming overlay instead.

This also brings back the original (-ish AFAIK) brightness of the pause menu, rather than it being much, much too light in 1.20.5+. The **one** downside of this whole commit is that the "Saving world" screen double-overlays dimming textures, making it _really_ dark. I'll see if I can fix it but probably not.