Fix Chest model error
Technically an optimisation, because this removes geometry that was previously rendered, albeit with a transparent texture.

This issue comes about as a result of FastChest's conversion of the Chest models internally to blocks rather than tile entities. To a client with FastChest, this fixes the 3D latch texture being present despite the chest texture having no pixels there, and to a Vanilla game, it just optimises Chest geometry. (Therefore this adds zero dependence on FastChest either)