Less blue PA Phantom Membrane
For some reason I had the impression the PA membrane texture wasn't in the in-game PA pack, but it totally is…

Anyway, main thing is that I added the PA texture in anyway, because I didn't like the bluish-purple tint the shading had. Some curves and reduction in blue levels for the real dark parts and hocus pocus it looks okay now.